Improvement: generated Batbridge code is now predicate descriptive
Instructions for Batbridge are defined in terms of a sequence of
terms, being either numbers or predicates. If a number is provided,
then when the appropriate assemble function is invoked that slot in
the instruction will be filled with the provided constant no questions
asked. Predicates allow for the placement of bounds on the remaining
fields of the instruction thus allowing guards for membership in some
range or set.