Improvement: Tweak to create a "batbridge" record
In the long term I want an entire ISA to come down to a single map
object which can be used by toothpick.core in a standard way to
assemble, encode and decode bytecode formats. This update is a first
step in that direction because it adds all the requisite map structure
for an O(N) serializer and deserializer.